requirements.txt
py
exif
fastapi
fastapi-cli
google-generativeai
piexif
pydantic
pydantic_core
requests
